If many new messages appear, allow syncFolderUIDRange to sync only 5000 [see desc]
This is ok because the next deep scan will pick up whatever messages we failed to retrieve. New messages will appear fast because a shallow scan will see them. Old messages will backfill at a rate of 10,000 per hour.

Maybe make it immediately re-scan if we detect we’ve bailed out, or keep a queue of UIDs we deferred?